feat: Provides portal context at root level
