+++ lib/components/Portal/Portal.tsx
@@ -1,34 +1,20 @@
-import React, { FunctionComponent, ReactNode, useContext } from 'react';
+import { invariant } from '@autoguru/utilities';
+import { FunctionComponent, ReactNode } from 'react';
 import { createPortal } from 'react-dom';
 
-import { PortalContext } from './context';
+import { usePortalContext } from './PortalProvider';
 
 interface Props {
 	children: ReactNode;
 	className?: string;
 }
 
-const buildPortalElement = (): Element => {
-	const element: Element = document.createElement('div');
-	element.setAttribute('id', `od-portal-${Math.round(Math.random() * 1e6)}`);
-	document.body.append(element);
-
-	return element;
-};
-
 export const Portal: FunctionComponent<Props> = ({ children }) => {
 	if (typeof window === 'undefined') return null;
 
 	// eslint-disable-next-line react-hooks/rules-of-hooks
-	const portalContext = useContext(PortalContext);
-	console.log(portalContext);
-	const context = portalContext || {
-		portalInstance: buildPortalElement(),
-	};
+	const portalContext = usePortalContext();
+	invariant(portalContext, 'There is no portal context provided');
 
-	return (
-		<PortalContext.Provider value={context}>
-			{createPortal(children, context.portalInstance)}
-		</PortalContext.Provider>
-	);
+	return createPortal(children, portalContext.portalInstance);
 };
