--- lib/components/Portal/context.ts
@@ -1,5 +1,0 @@
-import { Context, createContext } from 'react';
-
-export const PortalContext: Context<{
-	portalInstance: Element;
-}> = createContext(null);
